photo-urn/worker/requirements.txt
torch
torchvision
numpy
Pillow
opencv-python
trimesh
scipy
